requirements.txt
# ── AI / Extraction ──────────────────────────
anthropic>=0.40.0
groq>=0.9.0

# ── PDF / Image ──────────────────────────────
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10

# ── YOLO (layout detection) ──────────────────
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

# ── DXF Generation ───────────────────────────
ezdxf>=1.3.0

# ── API ──────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# ── Database ─────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# ── Utilities ────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0